web/satellite: dashboard lag message tooltip
When a user views their dashboard after uploading or downloading an object, we want to express the lag between uploading an object and when the metainfo loop runs, updating the satellite. This is so that they understand dashboard stats could be off a bit. This message should function as a tooltip and exist text to the dashboard title