pkg/quic,pkg/server: enable QUIC on server based on udp receive buffer size
Experiments have shown that QUIC transfers on high-bandhwidth connections
can be limited by the size of the UDP receive buffer. This buffer holds
packets that have been received by the kernel, but not yet read by the
application (quic-go in this case). Once this buffer fills up, the kernel
will drop any new incoming packet.

From benchmarking, we discovered that without increasing the receive
buffer, `quic-go` performs worse than `tcp+tls`. Therefore, if we are
not able to meet the required udp receive buffer size, we will disable
quic. This will be the default behavior.

User can override such default behavior by changing the config value of
`DisableQUIC` from `auto` to `true/false`. When the config value is set
to true or false, QUIC will be disabled/enabled regardless of the udp
receive buffer size on the host machine

Change-Id: I6875ad4cdffeb8510a83bbaa71cc2fd8834bb2be